+++ widget/src/main/java/com/jaspersoft/android/sdk/widget/report/view/ReportView.java
@@ -49,7 +49,7 @@ public class ReportView extends FrameLayout implements ReportWidget {
     }
 
     @Override
-    public void init(AuthorizedClient client, ServerInfo serverInfo, float scale) {
+    public void init(AuthorizedClient client, ServerInfo serverInfo, double scale) {
         reportViewerDelegate.init(client, serverInfo, scale);
         reportViewerDelegate.registerReportRendererCallback();
     }
